Reconstruct the project structure & Fixed spotless
