--- Iris-core/src/main/java/io/github/qylh/iris/core/annotation/Api.java
@@ -1,12 +1,0 @@
-package io.github.qylh.iris.core.annotation;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target(ElementType.METHOD)
-@Retention(RetentionPolicy.RUNTIME)
-public @interface Api {
-    String name();
-}
