--- Iris-core/src/main/java/io/github/qylh/iris/core/annotation/IrisRPC.java
@@ -1,11 +1,0 @@
-package io.github.qylh.iris.core.annotation;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target({ElementType.FIELD, ElementType.TYPE})
-@Retention(RetentionPolicy.RUNTIME)
-public @interface IrisRPC {
-}
