--- Iris-core/src/main/java/io/github/qylh/iris/core/annotation/IrisService.java
@@ -1,12 +1,0 @@
-package io.github.qylh.iris.core.annotation;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target(ElementType.TYPE)
-@Retention(RetentionPolicy.RUNTIME)
-public @interface IrisService {
-    String name();
-}
