+++ Iris-core/src/main/java/io/github/qylh/iris/core/common/annotation/IrisRPC.java
@@ -16,15 +16,14 @@
  *    specific language governing permissions and limitations
  *    under the License.
  */
-package io.github.qylh.iris.core.execption;
+package io.github.qylh.iris.core.common.annotation;
 
-public class MqttClientException extends Exception {
-    
-    public MqttClientException(String message) {
-        super(message);
-    }
-    
-    public MqttClientException(String message, Throwable cause) {
-        super(message, cause);
-    }
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target({ElementType.FIELD, ElementType.TYPE})
+@Retention(RetentionPolicy.RUNTIME)
+public @interface IrisRPC {
 }
