+++ Iris-core/src/main/java/io/github/qylh/iris/core/common/constant/Constants.java
@@ -16,11 +16,24 @@
  *    specific language governing permissions and limitations
  *    under the License.
  */
-package io.github.qylh.iris.core.listener;
+package io.github.qylh.iris.core.common.constant;
 
-import io.github.qylh.iris.core.common.msg.MqttMsg;
-
-public interface MqttMsgListener {
+public class Constants {
     
-    void onMessage(String topic, MqttMsg message);
+    public static final String MQTT_REQUEST_TOPIC_SUFFIX = "iris/request/";
+    
+    public static final String MQTT_RESPONSE_TOPIC_SUFFIX = "iris/response/";
+    
+    public static final String MQTT_CHARSET_NAME = "UTF-8";
+    
+    public static final int IRIS_MQTT_SUCCESS = 0;
+    
+    public static final int IRIS_MQTT_SERVICE_NOT_FOUND = 1;
+    
+    public static final int IRIS_MQTT_METHOD_NOT_FOUND = 2;
+    
+    public static final int IRIS_MQTT_INVOKE_ERROR = 3;
+    
+    public static final int QOS = 1;
+    
 }
