+++ Iris-core/src/main/java/io/github/qylh/iris/core/common/serializer/JacksonSerializer.java
@@ -16,7 +16,7 @@
  *    specific language governing permissions and limitations
  *    under the License.
  */
-package io.github.qylh.iris.core.serializer;
+package io.github.qylh.iris.core.common.serializer;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.core.type.TypeReference;
