+++ Iris-core/src/test/java/io/github/qylh/iris/core/serializer/JsonSerializerTest.java
@@ -18,6 +18,7 @@
  */
 package io.github.qylh.iris.core.serializer;
 
+import io.github.qylh.iris.core.common.serializer.JsonSerializer;
 import lombok.Data;
 import org.junit.Assert;
 import org.junit.Test;
