+++ Iris-spring-boot-starter/src/main/java/io/github/qylh/iris/spring/boot/IrisSpringBootStarterConfiguration.java
@@ -1,9 +1,27 @@
+/*
+ *    Licensed to the Apache Software Foundation (ASF) under one
+ *    or more contributor license agreements.  See the NOTICE file
+ *    distributed with this work for additional information
+ *    regarding copyright ownership.  The ASF licenses this file
+ *    to you under the Apache License, Version 2.0 (the
+ *    "License"); you may not use this file except in compliance
+ *    with the License.  You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing,
+ *    software distributed under the License is distributed on an
+ *    "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ *    KIND, either express or implied.  See the License for the
+ *    specific language governing permissions and limitations
+ *    under the License.
+ */
 package io.github.qylh.iris.spring.boot;
 
-import io.github.qylh.iris.core.annotation.IrisService;
+import io.github.qylh.iris.core.common.annotation.IrisService;
 import io.github.qylh.iris.core.config.MqttConnectionConfig;
-import io.github.qylh.iris.core.execption.MqttClientException;
-import io.github.qylh.iris.server.RequestProcessor;
+import io.github.qylh.iris.core.common.execption.MqttClientException;
+import io.github.qylh.iris.core.server.RequestProcessor;
 import org.springframework.beans.BeansException;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.ApplicationContext;
@@ -17,9 +35,9 @@ import java.util.stream.Collectors;
 @Configuration
 @EnableConfigurationProperties(IrisProperties.class)
 public class IrisSpringBootStarterConfiguration implements ApplicationContextAware {
-
+    
     private ApplicationContext applicationContext;
-
+    
     @Bean
     public RequestProcessor requestProcessor(IrisProperties properties) throws MqttClientException {
         MqttConnectionConfig mqttConnectionConfig = MqttConnectionConfig.builder()
@@ -39,15 +57,15 @@ public class IrisSpringBootStarterConfiguration implements ApplicationContextAwa
             }
             requestProcessor.putService(serviceName, entry.getValue());
         }
-        requestProcessor.start(serviceBeans.values().stream().map(Object::getClass).collect(Collectors.toList()) , mqttConnectionConfig);
+        requestProcessor.start(serviceBeans.values().stream().map(Object::getClass).collect(Collectors.toList()), mqttConnectionConfig);
         return requestProcessor;
     }
-
+    
     @Bean
-    public  IrisReferenceBeanPostProcessor irisReferenceBeanPostProcessor(IrisProperties irisProperties) {
+    public IrisReferenceBeanPostProcessor irisReferenceBeanPostProcessor(IrisProperties irisProperties) {
         return new IrisReferenceBeanPostProcessor(irisProperties);
     }
-
+    
     @Override
     public void setApplicationContext(ApplicationContext applicationContext) throws BeansException {
         this.applicationContext = applicationContext;
